All tests pass now, although the lack of service support on os x means that i have now disabled services on it
git-svn-id: https://reductivelabs.com/svn/puppet/trunk@789 980ebf18-57e1-0310-9a29-db15c13687c0

+++ lib/puppet/util.rb
@@ -98,7 +98,11 @@ module Util
         rescue
             raise
         ensure
-            File.delete(lock)
+            # I don't really understand how the lock file could disappear,
+            # but just in case...
+            if FileTest.exists?(lock)
+                File.delete(lock)
+            end
         end
     end
